test: 提取共享 fixture 到 conftest.py
